Added UI to indicate a new version
